pdf2md/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2md"
version = "0.1.0"
description = "Convert PDF documents into clean, structured, token-efficient Markdown for downstream LLM consumption."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "pdf2md contributors"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
marker = ["marker-pdf>=1.0.0"]
docling = ["docling>=2.50.0"]
all = ["marker-pdf>=1.0.0", "docling>=2.50.0"]
light = []
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
pdf2md = "pdf2md.cli:main"

[tool.setuptools.packages.find]
include = ["pdf2md*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
